Created a new module for output of land ice statistics
The new module is called li_statistics, in file mpas_li_statistics.F.
It is based on the glide_diagnostics module in CISM.
At time intervals specified by the user, various global statistics are
written to the standard output log file. These statistics include
the total ice area, volume and internal energy, along with the max/min
values of various state variables (thickness, temperature, velocity)
and the cells/edges/levels where these max/mins reside.

In addition, the user can write diagnostic output for a single grid cell
specified in the config file. This output includes cell thickness,
bed topography, SMB, etc., along with vertical profiles of temperature
and normal velocity at each sigma level.

To support these diagnostics, I modified the broadcast routines and added
some new minloc/maxloc subroutines in ../framework/mpas_dmpar.F.
Since these changes are in the framework, I included them in a previous
commit of branch framework/dmpar_bcast_maxloc. This statistics branch
was checked out from that framework branch.

In the Registry I added several config variables:
   config_ice_specific_heat (used in energy diagnostics)
   config_seconds_per_year (used for unit conversions in diagnostics)
   config_stats_interval (time interval for writing statistics, in number of timesteps;
                     the default is 0, in which case no stats are written)
   config_write_stats_on_startup (if true, write stats on startup)

In addition, I added two 2D state variables to the Registry:
   surfaceTemperature = upper ice surface temperature
   basalTemperature = lower ice surface temperature

Previously the temperature was defined only at layer midpoints, in the tracer array.
Note that the surface and basal temperatures are not tracers (they are not advected
and are not associated with any internal energy).
This treatment is different from CISM, in which surface and basal temperature
are part of the 3D temperature array.

I modified subroutine mpas_core_run to call subroutine li_compute_statistics
at startup (if desired) and at the designated timestep interval.

I also modified the Makefile to build the new statistics module.

I verified that statistics are written out at the desired interval in the
desired text format, which is similar to that of CISM log files and hopefully
is easy to read.

I also verified that the answers are the same (within roundoff) for 1, 2 or 4 processors
on a Mac. This confirms that the various global reductions are working as intended.