Adding input generation programs
These programs parse a registry file with some key/value pairs, and
generate a default namelist or a default streams file based on the
attributes associated with the namelist options, namelist records,
streams, or stream mebers.

For example:
    namelist_gen Reg.xml nl.input mode=initialize test=sw1
    streams_gen Reg.xml st.input mode=initialize test=sw1

Would write a default namelist and streams file that only had attributes
that had the attribute mode, and within that had the value initialize,
and also had the attribtue test which contained the value sw1.

When listing attributes within the Registry file, an arbitrary number of
values can be lised, separated with semicolons. For example:

mode="initialize;run;analyze"
test="sw1;sw2;sw3"

Would match the previous usage example.