Calling input generation targets
In this commit, namelist files and streams files are no longer generated
from registry. Instead they are generated by a core via the
core_namelist and core_streams targets that each core defines.

An arbitary number of namelist and streams files can be generated within
each model build.

All namelist and streams files that begin with the
name namelist or streams, and are placed within the inc directory are
copied into the "default_inputs" directory at the top level. Any file
that also doesn't exist in the top level directory, is copied there.