LI: remove inputHigherOrderVelocity stream
I've retained these variables being associated with a package, but
instead of creating a separate stream for them, they are included in the
input stream.  Because the package enables/disables the variables, MPAS
will only attempt to read them if the package is enabled.  Thus, having
a separate stream for them seems unnecessarily complicated.  This commit
and the previous one can be used as a reference if having multiple input
streams is ever desired.