Treat situation where point is not in any triangles in mpas_calculate_barycentric_weights_for_points
It is possible that a provided point (even a vertex) is not within any
valid triangle formed by 3 cell centers.  These changes allow
mpas_calculate_barycentric_weights_for_points to check for that
situation, and when it occurs the triangle that is used is the one
closest to the desired point (resulting in barycentric extrapolation,
but the extrapolation is done from the closest triangle).

In order to do this, I've added additional new routines:
mpas_distance_plane, mpas_distance

I've also provided more detailed documentation of the routines.