Fixing a bug related to indices in mpas_ocn_diagnostics.F
bug fix. in two places where the surface-average normal velocity is computed
the values were normalized with the wrong index: layerThicknessEdge(k,iCell)
was used when it should be  layerThicknessEdge(k,iEdge).

in cases tested so far, this led to  the surface-average normal velocity
to be zero and resulted in the ocean boundary layer being somewhat too shallow.