Moving the write for the restart timestamp
Previously, the restart timestamp file was written before writting the
actually restart file. If the run died while writing the restart file,
the restart timestamp file would not be incompatible with the run, as it
pointed to a non-existent time.

This commit fixes the issue by moving the write of the timestamp after
the write of the restart file.

Fixes: #331