Adding a default_value attribute to field types
This attribute is added in preparation for separating the bootstrapping
into two parts. The first bootstrapping part will setup the fields (but
not allocate them) and will be generated from the registry file. The
second part will allocate fields, and give them a default value, but
won't have access to registry fields.

This attribute is added to allow the default value to be setup after the
fields are allocated.