add CVMix/KPP option to fix ocean boundary layer depth
two config options are added. one logical to overwrite the KPP
computation of OBL, the second to specify the value used to
overwrite with.

this option will be useful idealized or testing configurations
where enhanced mixing is desired near the surface, but bounds
on the depth of that mixing is also needed.