Fixing an issue when generating nested var_structs
Previously, the name of a substruct was not extracted properly before
generating the call to generate a substruct of another struct. This
resulted in the call to the subroutine having (null) in it's name.

This commit fixes the issue by properly extracting the substruct name
when generating the subroutien call.