Always perform computations in mpas_dmpar_get_index_range() in double precision.
In dividing up an index space between tasks, we represent the number of indices in
the mesh, as well as the number of MPI ranks, as floating-point values. For very
large meshes, the number of indices may be so large as to not be representable exactly
as a single-precision floating-point number leading to errors in the divisiion of the
index space.