Development of mapper_SFo2g structure.
This commit results in a building executable, in which the mapper_SFo2g
structure is generated with a call to seq_map_init_rcfile.  Currently
the executable fails here because I haven't yet defined the appropriate
input mapping file in seq_maps.rc (or made the mapping file for that
matter).