Time averaging analysis member.
This is a new analysis member that does time reductions (averages, mins,
and maxes) over any arbitrary field in the Registry.

The temporal windows for reduction are in the name list: accumulation
duration, when to start accumulation, when to reset accumulation buffers,
and when to repeat accumulation. (It supports many different modes, such
as daily, monthly, yearly, seasonally, climatology, etc. through time
window configuration.)

The fields to be "averaged" are listed in the stream, and the analysis
member will modify the stream to produce averaged versions of those fields.

It is mostly feature complete, and there are three additions to be made
before it can be merged into ocean/develop:
1. restart capability (don't know how to do this yet)
2. ability to add the mesh to the stream (add_stream_fields didn't seem to work
   correctly for me, needs help.)
3. replication of the analysis member to allow for more than one stream
   (originally, the analysis member allowed for multiple stream, but
    the way the analysis member driver is written, I had to revert the
    analysis member to supporting only one stream. This means we need
    to duplicate the analysis member (macros? cut-and-paste?) a few times.)