LI: modified algorithm to compute HO normal velocity on edges
The function get_prism_velocity_on_FEdges now computes the normal
velocity on the edges midpoints of MPAS 2d cells, instead of computing
the averaged normal velocity on the edges as done before.
This algorithm assumes that the midpoint of cell edges belongs to one
of the triangles sharing the dual of that edge.
This assumption is less restrictive than the one needed in previous implementation.
The new algorithm has the same order of accuracy of the previous algorithm
even if the fluxes through the edge computed using the normal velocity
are not anymore exact for linear/bilinear prismatic finite elements.