Updating default number of threads
Mira node can run 64 tasks: 16 cores x 4 threads/core. Login nodes can run 48 tasks. This commit increases the number of concurrent builds from 4 to 8 to speed up the builds. Also it updates the default number of threads for new cases from 12/24 to 16 to avoid 'not a power of two' errors and to reflect the node architecture. The namelist variable values for PEs are changed.

[NML]