Add halo exchanges in new time-integration scheme to ensure bit-restartability.
We need to exchange halos for theta_m (time level 2), pressure_p, and rtheta_p
at the end of all but the last dynamics substep.

It's clear that these exchanges are needed to recover the correct behavior of
the old time integration scheme, but it's still not clear why these are needed
for restartability.