Adding missing scratch root for blues
CESMSCRATCHROOT is not defined in the blues configuration. So
users need to specify "-sharedlibroot" when trying to build
and run tests using create_test. If the user misses the
"-sharedlibroot" argument the build fails. This commit fixes
this issue by explicitly specifying the scratch root for blues.

Also see Issue #154

SEG-29