Make GLC grid independent of LND grid
Previously it was required that the GLC grid be the same as the LND grid
from the perspective of the coupler.  However, for ACME we require a
full-fledged GLC grid.  These changes to scripts and model code
eliminate the way the GLC grid was set to the LND grid and make the GLC
grid a fully-fledged grid.  As part of this, the CISM_GRID
information has been eliminated.

I have not updated existing compsets that use CISM, so CISM will not run
until they are updated.