Adding namelist defaults for CONUS VarRes grid
These changes allow a user to build the variable resolution CONUS RRM
ACME configuration out-of-the-box without having to specify the values
for some physics paramterizations in the user_nl_* files.   WARNING:
As of this commit, the values have not been tuned for the CONUS grid but
are values used to tune either the ne30 or armx8v3 grid.

 On branch eroesler/ATM/RRM-meshfilefixer
 Changes to be committed:
	modified:   namelist_files/namelist_defaults_cam.xml