LI: Fix logic to only update FEM mesh when dynamic vertex mask changes
Currently the FEM mesh would be updated whenever the vertexMask changes,
but it is a bitmask with a number of different components.  I've changed the
logic to only check the bit for dynamic ice extent and not the other bits.  To
reflect this change, I've also changed the name of the variable used for this
from anyVertexMaskChanged to anyDynamicVertexMaskChanged.