Fixing an issue with mpas_decomp that prevents building on macs
This commits changes the abstract interface for a decomposition method.
Previously the result frmo the decomposition method function was the
owned indicies array pointer, but gfortran built on macs using macports
(it's unclear if all of these are needed or not) fails to build the old
version.

The new abstract interface changes ownedIndices to be in the input
argument list, and changes the output result to be the iErr flag. Other
than that, the decomposition methods work the same as they did before.