Update to work with mpi-serial2.0
TryMPISERIAL.f90 is supposed to fail to compile with mpi-serial to let
the PIO configure (with CMake) know that it should set its build for mpi-serial.
The old version was compiling successfuly with the previous mpi-serial.
New version now fails to compile with mpi-serial.

Tested it by hand to make sure it compiles with openmpi but not
mpi-serial.

SEG-50