LI: Allow restarts for HO dycores
This required adding some new fields as restart variables: beta, uReconstructX/Y

It also required reorganizing the logic for when the FEM mesh gets generated. To
ensure BFB restarts, the FEM mesh is now generated for the first time in
li_velocity_external_block_init.  This ensures the mesh will always be created.
When li_velocity_external_solve is called, there still is a check for if the dynamic
ice extent has changed, in which case the FEM mesh is re-generated.
However, I've changed the vertexMask initialization so that
on init the vertexMask will appear to be unchanged.  On a cold start this is fine,
because the FEM mesh was generated during init.  On a restart, velocity is NOT solved
on init, but the FEM mesh is still generated on init - this is necessary for the solver
to work correctly on the first time step if the dynamic ice extent has not changed at
that time.  (Note that the FEM mesh will be re-generated on the first time step
if the dynamic ice extent has changed.  This will be an unnecessary
operation, but accounting for this case would add substantially to the logic.)