Combining forward and analysis modes of the ocean core
This commit combines the forward and analysis build modes of the ocean
core into a single executable that is controllable via namelist options.

To build the model now, you can use:
    make target CORE=ocean

This creates ocean_model. In the namelist file, there is an option named
"config_run_mode" that controls the run_mode for the executable (setting
it to 'forward' will cause the executable to run in forward mode, while
a value of 'analysis' will cause the executable to run in the analysis
mode).