wait_for_tests: In CDash, segregate namelist failures from real failures
Frequent namelist diffs hid a real BFB change in our tests. This
commit assigns tests that only had a namelist change to the 'notrun'
status. This is a bit confusing, but it's the only other option than
pass or fail.

Also, change doctest unit tests to always be verbose.

[BFB]

SEG-92