Make output for acme python scripts unbuffered
Interaction with python print statements and output from shell
commands was not working as intended. In many cases, the python output
was buffered and then a shell command with a lot of output was executed
resulting in a long delay before the python prints would appear.
In some cases, this made it harder to find/debug problems.

[BFB]

SEG-87