correct surface focing
this commit corrects two bugs in the surface forcing.

the first is a soft bug, corrected by changing transmission_coefficients to fractionAbsorbed:
we distribute the non-penetrative fluxes in the vertical by computing the transmission coefficient
at layer interaces. The difference between the transmission coefficients at top and bottom of the cell
is the fraction of the surface flux absorbed in that cell. Previously we called this amount
the "transmissionCoefficient" but in reality it is the fractionAbsorbed. This name change is made throughout.

the second is a hard bug, corrected by the new computation of the temperature tendency due to shortwave forcing:
shortwave comes from the coupler in units of W/m2. We multiply by "hflux" to change the units to K m / s.
we accumulate the temperature tendency in tend_tracer which has units of K m / s, where "m" is from the layer thickness h.

tend_tracer =   d/dz ( h S_{interface}) where S_{interface} is the flux at the top and bottom of the layer interface.

expanding, we have

tend_tracer(k) = ( h(k) * (S(top) - S(bottom) / h(k) ) = S(top) - S(bottom)

S(top) = penetrativeTemperatureFlux * weight(top)
S(bottom) = penetrativeTemperatureFlux * weight(bottom)

so

tend_tracer(k) = penetrativeTemperatureFlux * (weight(top)-weight(bottom))

previously we were dividing by layer thickness on the RHS of the line above. This was incorrect.