Adding core specific framework routines modules
This commit adds (as of yet, unused) framework routines modules in
preparation for generalizing the MPAS framework.

These modules will house all of the #include's a core will use, in
addition to providing routines to the subdrive to setup function
pointers, which will be used throughout framework.

In addition, a core_abbrev attribute is added to the registry block in
Registry.xml file. This allows the parsing tool to generate functions
and subroutines using the core selected abbreviation.