LI: Adjust HO restart fix
The implementation in the previous commit did not work correctly because
it was setting the FEM mesh before vertexMask was calculated for the
first time.  This is an alternate implementation that goes back to
setting the FEM mesh only in the velocity_solve routine (i.e., I've
removed the set of FEM mesh from the velocity_block_init routine).  At
the end of the initial time, I set vertexMask to 0 if the run is a
restart with the HO-dycore.  This will ensure the FEM mask gets created
on the first time step (since the velocity solver is not called on init
with a HO dycore).