Fix underscores in okubo weiss c code
This commit fixes an issue related to c / fortran interoperability
due to underscores. It mainly updates the Okubo Weiss c code to follow
the rest of the MPAS format related to c code (defining the function
names based on how many underscores, if any, they should have).

In addition, it forces c code in the analysis member directory to obey
the CFLAGS definition.