Adding a TOOL_TARGET_ARCH option
This commit adds a new build option named TOOL_TARGET_ARCH which will
pass a build flag to the compiler when building all of our tools
(registry, input generators). This is required for machines like edison,
where executables are not execuatable on login nodes, but tools like
registry are required to be executable on login nodes.