Modify buildlib.mct for mpi-serial2.0 build.
For builds with the mpi-serial library....
The mpif.h has to be copied from the source to the $LIBDIR/include before
MCT configure is invoked so that mpeu will find it.  Also
copy the mpi-serial/*.h files to $LIBDIR/mct/mpi-serial after the build
because gptl expects them to be there.

[BFB]

SEG-50