Allow immutable streams not in streams file to be added to other streams
Currently, if an immutable stream is specified as a member of another stream
that is listed in a streams.CORE file, the immutable stream will only be added
if it is listed in the streams file itself, e.g.:
   <immutable_stream name="basicmesh" type="none" filename_template="not-to-be-used.nc"/>
This is a bit redundant, because all immutable_streams are defined in Registry.xml,
so it would be possible to eliminate the specification of these immutable_streams
within streams.CORE files.

This commit introduces changes to eliminate this restriction.  I.e., now any
immutable_streams defined in Registry.xml can be included as members of other
streams in streams.CORE files without needing to list the immutable_stream
separately in the streams.CORE file.

This is implemented by adding a new optional argument to subroutine
MPAS_stream_mgr_add_stream_fields that is called only_add_if_immutable.  If this
is true and the stream is mutable, then the routine returns an error.

The other piece is that xml_stream_parser has been modified to use this new
optional argument when it calls MPAS_stream_mgr_add_stream_fields.
Specifically, when xml_stream_parser attempts to add other streams to a stream,
it first calls MPAS_stream_mgr_add_stream_fields with only_add_if_immutable=.true.
If an error is returned, then xml_stream_parser attempts to add a mutable stream.