LI: Reorganizing time levels
Previously MPAS-LI had 2 time levels for all variables and used the
convention that time level 2 was the new time level (to be solved)
and time level 1 was the old time level.

With the var_struct reorg in the previous commits, almost all variables
now have 1 time level, and only a few variables that require a second
time level have 2.  Therefore, to minimize confusion, I have modified
the code to use a new convention that time level 1 is always the
*current* time to be solved, and time level 2, where present, is the old
value of a variable.  In other words, all calculations should be done on
time level 1, unless an explicit need for the old value of a variable is
needed.  As part of this, I have removed the timeLevel argument from all
subroutines - it should only ever be used explicitly to get a specific
time level (current or old).