Reduce memory usage in atm_advance_scalars routines.
The routines atm_advance_scalars() and atm_advance_scalars_mono() routines make
use of arrays rho_zz_int and scalar_tend_save (for the unlimited routine) when
re-integrating the density field and for storing the scalar tendency field from
the dynamics, but only when scalar transport is split from the dynamics RK steps.
These arrays were allocated on the stack, which led to unnecessary memory usage
(and an unecessary memory copy in the case of rho_zz_int) when scalar transport
is not split.

Now, these two arrays are allocated on the heap only when the routines are used
for split scalar transport; otherwise, the rho_zz_int pointer is simply set to
point to the rho_zz_new array.