Initial implementation of physics suite option in MPAS-Atmosphere.
This commit adds a new namelist option, config_physics_suite, that may
be used to identify an entire suite of physics schemes; the default
value for this option is currently 'wrf', and a 'none' suite is also
supported, which disables all physics schemes. The contents of each suite
are defined in src/core_atmosphere/physics/mpas_atmphys_control.F.

The default value for all individual physics scheme namelist options
is now 'suite', which instructs MPAS-Atmosphere to use whatever scheme
the selected suite provides. However, it is possible to override any
of the schemes by specifying them explicitly in the namelist.