Add baseline_gen_comp script
The baseline_gen_comp script has the functionality of the old
component_gen_comp. However, now it is a wrapper to Mariana's
component_compgen_baseline.sh. This means that it compares/generates for ALL
components at once.

This is meant to be run after all tests in a given test suite are complete. Some
typical uses are: (1) re-comparing against baselines if some baselines did not
yet exist when the test suite was run, (2) comparing against a different set of
baselines from the ones originally specified when you created the test suite,
(3) generating baselines after the fact (if you ran the test suite without
-generate).

Also made some minor changes in component_compgen_baseline.sh:
- Added an optional -cprnc_exe argument, so that it can be run without giving a
  CASEROOT
- Fixed documentation, and added more error-checking of arguments