First commit of ice-sheet-to-ocean coupling development work.
This commit is first in branch to develop ice-sheet-to-ocean coupling.
A note: this was branched off of branch
jeremyfyke/cpl/ice_ocean_coupling, at the point where a PR was requested
to merge this branch into master.  Thus, it is expected that one or more
rebase operations will be required to the current branch, once the
jeremyfyke/cpl/ice_ocean_coupling branch is merged.

Developments within this commit: addition of a new glc to coupler field
(Sg_topg), which is a test field that is meant to pass ice sheet bedrock
topography to the ocean.  This field is created in the g2x attribute
vector, filled, and exported.

Build-tested, not run-tested.