Machine file fixes for Cray compiler on Titan
CMakeLists.txt for PIO is altered to enforce -O 0 when the Cray compiler is being used. This affects all machines using Cray compilers, so I'm asking someone else to integrate this so they can take a look. Eventually, I will go back to this and do a more selective downgrade of optimizations for PIO.

The GPTL timing Makefile had to be altered to assume a default of "false" for the threading variable. This is because the Cray compiler requires it to be explicitly set to false in order to put "-h nomp" since Cray defaults to using OpenMP.

The Depends.titan.cray file lowers optimizations moderately for derivative_mod.F90, only turning off the automagical fusion Cray likes to use. Optimizations are removed entirely for an aerosol file.

The config_compilers.xml file is altered for more approprite flags for Cray only when run on Titan.

Finally, the env_mach_specific.titan file uses the most recent cray compiler.

With these modifications, ACME builds and runs successfully on Titan with the Cray compilers for a B1850C5 compset on an ne30_g16 grid.