LI: Have diffusivity use uReconstructZonal/Meridional
This eliminates the restriction that this can only be computed on a
plane.  This takes advantage of the fact that mpas_reconstruct assigns
uReconstructZonal/uReconstructMeridional equal to uReconstructX/Y for
planar meshes.  I.e, uReconstructZonal/uReconstructMeridional always
contain orthogonal components of the vector in the plane of interest,
whereas uReconstructX/Y will only be applicable to planar meshes.  So
using uReconstructZonal/uReconstructMeridional eliminates the need to
include logic for sphere vs. plane.