LI: Add config_adaptive_timestep_force_interval option
This allow the user to specify an interval at which the clock should hit
when using the adaptive time stepper.  This can be used to ensure that the
model would output at specified intervals, or for coupling to a climate
model to ensure that the model will run for exactly the coupling
interval and not longer.