Always return 'input' as the name of the mesh stream for the test core
Previous logic in test_get_mesh_stream() checked on the non-existent namelist
option config_do_retart; if this option was not found, the function returned
a blank stream name. Since the test core doesn't even have a restart stream,
it seems simpler for now to always use the only input stream defined in the
Registry.xml file as the mesh stream.