Set default config_stop_time for test_core to the same as config_start_time
Both config_stop_time and config_run_duration were previously set to "none",
which caused a run-time error in the test_setup_clock function when attempting
to use config_run_duration as the timestep.

Now, we set config_stop_time to the same time as config_start_time by default
in the Registry.xml file, and we set the timestep to
(config_stop_time - config_start_time) if config_run_duration is "none"; however,
if config_run_duration is not "none", it is used as the timestep instead.