LI: Zero velo on uphill margin edges
Don't allow normalVelocity on edges where an unglaciated cell with
higher elevation neighbors a glaciated cell.  Some velocity solvers
could generate a nonzero velocity on these edges.  In the case of a
velocity directed into the ice sheet, this probably does no harm
for advection because there is no ice to advect in, but it could result
in overly restrictive advective CFL conditions.  In the case of velocity
directed out of the ice sheet, this would result in uphill flow which is
highly unlikely to be physically correct.  (It could be possible in a HO
stress balance where stress transfer 'overrides' the driving stress, but
this seems unlikely to be significant.)  Therefore, always zero velocity
in these situations.