Bug fix for Consine Solar Zenith Angle calculation
This commit fixes consine solar zenith angle calculation
so that it is averaged over a radiation time step.
Currently, the consine solar zenith angle is held
constant during a radiation time step, which is a bug.
This commit introduces a flag (sol_insolation_bug_fix)
which when turned true fixes this bug. This flag is .false.
by default as turning this true is an answer changing
change (not climate changing). See Issue #139 for more
details.

This commit adds a logical flag to the cam_inparm block of
atm_in namelist

Fixes #139
[NML]
AG-189