Use specified io_type when creating streams in stream manager
The mpas_stream_manager code previously used an io_type of MPAS_IO_PNETCDF
in all calls to MPAS_createStream. With this commit, MPAS_IO_PNETCDF is replaced
with the io_type that was specified for the stream, either through the call
to MPAS_stream_mgr_create_stream() or MPAS_stream_mgr_set_property().