Use io_type associated with mesh stream during bootstrapping
During the bootstrapping procedure, there are two points where we open input
files using the low-level MPAS I/O layer via the subroutine MPAS_io_open(),
and at which time we need to provide an io_type for the file to be read.

In order to obtain the correct io_type in mpas_bootstrap_framework_phase1(), we
add a new output argument to the xml_stream_get_attribute() routine that reads
the 'io_type' attribute from the mesh stream in the run-time XML stream
configuration file. This io_type is returned as a string and converted to
an integer in mpas_init(), where it can then be passed to
mpas_bootstrap_framework_phase1().

In mpas_bootstrap_framework_phase2(), we read dimensions from all active input
streams using the low-level MPAS I/O layer; however, at this point in the code
execution, streams have been setup, and we can simply query
the MPAS_STREAM_PROPERTY_IOTYPE property for each of the input streams to get
the io_type for use in the call to MPAS_io_open().