Change return type of get_mesh_stream() function to avoid runtime errors with ifort
The ifort 13.1.2 compiler appears to have problems with function pointers that
return character strings. To work around this issue, we change the return type
of the mpas_get_mesh_stream_function() interface to be an integer error code;
the name of the mesh stream that was formerly the return type is now returned
via an intent(out) argument.

After changing the interface of the get_mesh_stream() function pointer in
the core_type type, we also update the mesh stream function implementations
in all cores.