Fixes to get split transport working properly with OpenMP code
With this commit, the code now provides bit-identical results with or
without physics (see note below!), with either the old integration scheme
or the split-transport scheme for any combination of MPI task count and
OpenMP thread count.

NB: The results are only reproducible with the Kain-Fritsch scheme; for yet
unknown reasons, we're getting different results with the Tiedtke scheme.