Avoid writing initial diagnostics output file on restart runs
Some diagnostic fields are not computed at model startup, and therefore aren't
available in the initial diagnostics file. For a restart run, it's reasonable
to expect that a diagnostics file would have been written at the same valid time
in the preceding segment of the run and would include valid diagnostics. Consequently,
we should follow what is already done for the history stream and not write the initial
diagnostics output file on model restart.

If it were ever necessary to get an initial diagnostics output file on model restart,
one could simply define a new stream that contains all diagnostic fields and has
output_interval="initial_only".