Correctly detect non-isobaric data in real-data initialization
The code to check whether we likely were using isobaric initial condition
data was not correctly limiting the array bounds of its search for valid
pressures, both in the vertical and in the horizontal when running in
parallel.

Also, if model-level is detected (assumed, really), we explicitly fill
in the surface pressure and height fields, similar to what is done for
the surface pressure in isobaric data cases.

These changes allow us to initialize with native-grid GFS data converted
to intermediate format by the rd_sig program.