Allow nondynamic 'inlet' edge case for velo check for nondynamic edges
After external velocity solves return velocity, we check if the velocity solver has
returned a velocity on any non-dynamic edges.  We recently discovered an edge case where
this is ok so this commit allows it without error.

If there are two peninsulas of dynamic ice
with a single 'row' of nondynamic cells between them, the FEM velo solver will likely
calculate a nonzero velocity on an edge that has 0 thickness.  The two FEM elements
neighboring this edge have nonzero thickness everywhere except along this edge, and
so there is no guarantee of zero-velocity on this edge.  Schematically, this looks like:

    \   I   /
   A |--e--| A
    /   I   \

where the lines are edges, and e is the edge with the issue.  I's are inactive cells, and
A's are active cells.  So check for this specific situation before calling this an error.

Additionally, set the edge velocity to 0 on these edges.