Move CFL error check to after adaptive timestep update
The CFL check that sets an error if the timestep is too small
was occurring before the adaptive timestepper was setting the
new time step value.  In most cases this did not lead to an error
but could if the previous time step was longer than the current
CFL criteria.

This was also complicated by the fact that if the adaptive timestepper
was being used, the variable dtSeconds was uninitialized.  If the compiler
initialized it to 0, there was no problem, but in optimized mode, the
compiler I was testing was setting it to a large value, which triggered
a CFL error because the updated value for dtSeconds had yet to be applied.
I have also initialized the dtSeconds variable to 0 when the adaptive
time stepper is used to avoid an error from tend_layerThickness_fo_upwind.