Compute exact number of non-zeros in rof->ocn map
I coded up the algorithm I laid out in the pull request, and it worked for the
rx0.5 -> gx3v7 test. Bill, can you test it with your glacier grid? Note that
there is now a line of output telling you the estimated number of non-zeros in
B (and I left the namelist variable in place if there is a time when I
under-estimate the size, but I'm 99% sure this is exact).