Process input streams individually and write a message to the log for each stream that is read
The existing code didn't write messages of any sort to the log files when input
streams were read during model integration, and thus, users had no way of telling
from a log file whether, e.g., SST updates were actually taking place in the simulation.
Furthermore, because input streams aren't required to have input fields at the exact
times of the reads (MPAS_STREAM_LATEST_BEFORE is used), users would not have any
indication of the valid times that were actually being used in the simulation.

This commit replaces the blanket read of all input streams with a loop that considers
each input stream individually and writes a message including the valid time of
the read to the log file when a stream is read.