Set default surface update interval to 86400 s during initialization
Setting the default output_interval for the "surface" stream in the initialization
core to 86400 s, and correspondingly, the config_fg_interval, will require fewer
changes to files for anyone working with daily SST and sea-ice analyses.

There's also no harm in setting the output_interval for the "surface" stream
to a value other than "none", since this stream is only activated when creating
a surface update file.