Added crop harvest and yield calculator
These code modifications include yield calculation in bu/acre and
dry matter (t/ha) and allow harvest of grain to a product pool
analogous to the wood product pools, but with a 1-yr turnover.
An option to harvest percent of residue is allowed but currently
residue harvest is set to 0.

Three new parameters are in the clm_params file for the yield and
residue harvest. The new parameter file (clm_params.crop.c150330.nc)
will be used only when the crop model is active

[CC]

LG-71