Only write 'restart_timestamp' file from MPI rank 0
This file was previously written by all MPI tasks, which seems slightly
dangerous and potentially detrimental to performance for large task counts.

There appears to be no downside to this change: if an MPI rank >0 crashes
before reaching the point where the 'restart_timestamp' file is written, rank 0
will write the file (same behavior as before, though not necessarily desirable:
rank 0 will still write the file), and if MPI rank 0 crashes before reaching
the code to write the file, the file will not be written (which is desirable:
the file should ideally only be written if all tasks successfully get past
writing the restart file).