Ensure test results remain on Cdash dashboard for maximum length of time
CDash is set up to purge a result after 24 hours have passed since
the NightlyStartTime of the test, NOT it's completion time. Since we
are faking the XML anyway, let's just pretend our tests started and
finished at the same time. This should give us the full 24 hours on
our dashboard.

I still don't know what to do about the tests we have that run for
multiple days.

[BFB]