Use our native gamma/erf implementations on Intel.
Christopher Kerr found that, when using version 15.0.x of the Intel
Fortran compiler, the compiler's version is actually slower than the
implementations that exist in csm_share. This is at least in part
because Intel relies heavily on vectorization, but it doesn't inline,
and therefore can't auto-vectorize, these functions. Something similar
apparently holds for the Cray compiler.

Furthermore, we use "good enough" approximations that have the precision
needed for the ranges used in CESM (particularly for CAM microphysics),
and therefore can get away with using algorithms that are cheaper, but
less precise for large arguments.

Therefore, until Intel's implementation improves, we ignore the gamma
intrinsic and use our own.

Testing:

    Test suite: Unit tested only, using NAG (this code has been used
        with all supported compilers, however, so we have high
        confidence in it).
    Baseline: N/A
    Namelist changes: None.
    Status: roundoff (for all CAM/CLM configurations)

Code reviews:

     None so far.