Better and safer test cleanup
Leverage the testid capability for create_test to make it
easy to determine which files in a machine's testroot were generated
by jenkins. The previous implementation removed <testroot>/*machine*
which was dangerous as it could clobber any ongoing ACME testing
that was happening outside of Jenkins and also failed to remove old
sharedlibroot directories which were starting to build-up on our testing
machines.

[BFB]