Fix bug in HO velo check for no ice
Commit d219fac introduced a check for if the domain has no dynamic ice,
in which case external velocity solvers are not called.  This is because
external velocity solvers may abort if there is no ice.

However, the implementation was written in a way that would cause
problems (e.g. Albany hangs) if some procs have no dynamic ice but
others do.  This commit makes a more general implementation that
properly handles that situation.