Don't abort if errors occur when computing the MSLP diagnostic
The code to compute mean sea-level pressure looks for a layer that
is 100 hPa above the surface; if this layer was not found, the code
previously called mpas_dmpar_abort(). However, for certain idealized
test cases where a base-state pressure is not supplied (specifically,
the mountain wave test case), the derived full pressure field is not
valid, and the trap in the MSLP code would cause the model to abort.

Rather than aborting, it is much preferable to simply print a message
informing users that the diagnostic will not be computed, and to continue
with the simulation.