Move computation of invAreaCell field above first call to atm_compute_solve_diagnostics()
The atm_compute_solve_diagnostics() routine is called from the MPAS-A init routine,
and thereafter, toward the end of each timestep. This routine makes use of the invAreaCell
field, which wasn't computed until after the first call to the routine during model
init. Besides leading to some incorrect diagnostic fields in the first model timestep,
this broke bit-identical restarts of the model.